+++ packages/storey/src/containers/map/key/key_set.rs
@@ -1,0 +1,4 @@
+/// The default key set for use with a [`Map`](super::Map).
+///
+/// To find out more about key sets, take a look at the [`Key`] trait's documentation.
+pub struct DefaultKeySet;
